fix: Refactor value/list/object fields into separate files.
