+++ src/index.ts
@@ -1,11 +1,7 @@
-export {
-  createObjectState,
-  FieldState,
-  ListFieldState,
-  ObjectConfig,
-  ObjectState,
-  required,
-  Rule,
-} from "src/formState";
+export { ObjectConfig } from "src/config";
+export { ListFieldState } from "src/fields/listField";
+export { createObjectState, ObjectState } from "src/fields/objectField";
+export { FieldState } from "src/fields/valueField";
+export { required, Rule } from "src/rules";
 export { useFormState } from "src/useFormState";
 export { useFormStates } from "src/useFormStates";
